Move previewer.py into browser folder
